+++ presentation/src/main/res/values/strings.xml
@@ -5,4 +5,5 @@
     <string name="species">SPECIES: %1$s</string>
     <string name="gender">GENDER: %1$s</string>
     <string name="back">Back</string>
+    <string name="episodes">EPISODES</string>
 </resources>
